Refactor: [BADA-105] getTradeDetail.ts 분리 및 참조 경로 일원화 #59
