+++ src/app/trade/data/[id]/page.tsx
@@ -1,4 +1,4 @@
-import { getTradeDetail } from '@features/trade/apis/getTradeDetail';
+import { getTradePostDetail } from '@features/trade/apis/trade';
 import { TradeDetailPage } from '@features/trade/pages/TradeDetailPage';
 
 interface Props {
@@ -9,7 +9,7 @@ export default async function DataDetailPage(props: Props) {
   const { params } = await props;
   const { id } = await params;
   // 데이터 게시물 상세 API 호출
-  const { postUserId, post, sellerName } = await getTradeDetail(id);
+  const { postUserId, post, sellerName } = await getTradePostDetail(id);
 
   // 데이터 게시물 전용 UI만 구현
   return (
@@ -17,7 +17,7 @@ export default async function DataDetailPage(props: Props) {
       tradeId={id}
       post={post}
       postUserId={postUserId}
-      postType="data"
+      postType="DATA"
       sellerName={sellerName}
     />
   );
