+++ src/features/trade/models/post.ts
@@ -1,3 +1,4 @@
+// 전체 게시물 리스트 조회용
 export interface Post {
   id: number;
   title: string;
@@ -10,3 +11,33 @@ export interface Post {
   likesCount: number;
   isLiked: boolean;
 }
+
+// 게시물 상세 조회용
+export interface TradeDetailPost {
+  id: number;
+  title: string;
+  comment: string;
+  price: number;
+  deadLine: string;
+  postImage: string;
+  isSold: boolean;
+  createdAt: string;
+  postCategory: PostType;
+
+  // 데이터 거래용
+  mobileCarrier?: string;
+  capacity?: number;
+
+  // 기프티콘 거래용
+  issueDate?: string;
+  partner?: string;
+  likesCount: number;
+  isLiked: boolean;
+}
+
+export interface TradeWriter {
+  userId: number;
+  username: string;
+}
+
+export type PostType = 'DATA' | 'GIFTICON';
